requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
altair=5.5.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
attrs=25.3.0=pypi_0
blas=1.0=mkl
blinker=1.9.0=pypi_0
bottleneck=1.4.2=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_9
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
cachetools=5.5.2=pypi_0
click=8.1.8=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.1=py310hdb19cb5_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
debugpy=1.8.14=py310hf71b8c6_0
decorator=5.2.1=pyhd8ed1ab_0
exceptiongroup=1.2.2=pyhd8ed1ab_1
executing=2.1.0=pyhd8ed1ab_1
expat=2.7.1=h6a678d5_0
faiss-cpu=1.10.0=pypi_0
filelock=3.18.0=pypi_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.55.3=py310h5eee18b_0
freetype=2.13.3=h4a9f257_0
fsspec=2025.3.2=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.44=pypi_0
icu=73.1=h6a678d5_0
importlib-metadata=8.6.1=pyha770c72_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.35.0=pyh907856f_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
jpeg=9e=h5eee18b_3
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.8=py310h6a678d5_0
krb5=1.20.1=h143b758_1
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20250127.0=cxx17_h6a678d5_0
libcups=2.4.2=h2d74bed_1
libcurl=8.12.1=hc9e6f67_0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libglib=2.78.4=hdc74915_0
libgomp=14.2.0=h767d61c_2
libiconv=1.16=h5eee18b_3
libnghttp2=1.57.0=h2d74bed_0
libpng=1.6.39=h5eee18b_0
libpq=17.4=hdbd6064_0
libprotobuf=5.29.3=hc99497a_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.11.1=h251f7ec_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.7.0=hde9077f_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.7=hfdd30dd_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=3.0.2=pypi_0
matplotlib=3.10.0=py310h06a4308_0
matplotlib-base=3.10.0=py310hbfdbfaf_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_2
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
mpmath=1.3.0=pypi_0
mysql=8.4.0=h721767e_2
narwhals=1.35.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
numexpr=2.10.1=py310h3c60e43_0
numpy=2.2.4=pypi_0
numpy-base=2.0.1=py310hb5e798b_1
nvidia-cublas-cu12=12.4.5.8=pypi_0
nvidia-cuda-cupti-cu12=12.4.127=pypi_0
nvidia-cuda-nvrtc-cu12=12.4.127=pypi_0
nvidia-cuda-runtime-cu12=12.4.127=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.2.1.3=pypi_0
nvidia-curand-cu12=10.3.5.147=pypi_0
nvidia-cusolver-cu12=11.6.1.9=pypi_0
nvidia-cusparse-cu12=12.3.1.170=pypi_0
nvidia-cusparselt-cu12=0.6.2=pypi_0
nvidia-nccl-cu12=2.21.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu12=12.4.127=pypi_0
openjpeg=2.5.2=h0d4d230_1
openldap=2.6.4=h42fbc30_0
openssl=3.5.0=h7b32b05_0
packaging=24.2=py310h06a4308_0
pandas=2.2.3=py310h6a678d5_0
parso=0.8.4=pyhd8ed1ab_1
pcre2=10.42=hebb0a14_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.2.1=pypi_0
pip=25.0=py310h06a4308_0
platformdirs=4.3.7=pyh29332c3_0
prompt-toolkit=3.0.51=pyha770c72_0
protobuf=5.29.4=pypi_0
psutil=7.0.0=py310ha75aee5_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=19.0.1=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.0=py310h06a4308_0
pyqt=6.7.1=py310h6a678d5_1
pyqt6-sip=13.9.1=py310h5eee18b_1
python=3.10.16=he870216_1
python-dateutil=2.9.0post0=py310h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytz=2025.2=pypi_0
pyzmq=26.2.0=py310h71f11fc_1
qtbase=6.7.3=hdaa5aa8_0
qtdeclarative=6.7.3=h6a678d5_0
qtsvg=6.7.3=he621ea3_0
qttools=6.7.3=h80c7b02_0
qtwebchannel=6.7.3=h6a678d5_0
qtwebsockets=6.7.3=h6a678d5_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
rpds-py=0.24.0=pypi_0
seaborn=0.13.2=py310h06a4308_2
setuptools=75.8.0=py310h06a4308_0
sip=6.10.0=py310h6a678d5_0
six=1.17.0=py310h06a4308_0
smmap=5.0.2=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
streamlit=1.44.1=pypi_0
sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=9.1.2=pypi_0
tk=8.6.14=h39e8969_0
toml=0.10.2=pypi_0
tomli=2.0.1=py310h06a4308_0
torch=2.6.0=pypi_0
torchvision=0.21.0=pypi_0
tornado=6.4.2=py310h5eee18b_0
traitlets=5.14.3=pyhd8ed1ab_1
triton=3.2.0=pypi_0
typing_extensions=4.13.2=pyh29332c3_0
tzdata=2025.2=pypi_0
unicodedata2=15.1.0=py310h5eee18b_1
watchdog=6.0.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
wheel=0.45.1=py310h06a4308_0
xcb-util-cursor=0.1.4=h5eee18b_0
xz=5.6.4=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zipp=3.21.0=pyhd8ed1ab_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
